typecast for comparison, reformat comments

+++ 2-selection_sort.c
@@ -3,11 +3,6 @@
 
 void selection_sort(int *array, size_t size)
 {
-            
-        // search array for smallest element
-        // move to first search position
-        // find next smallest, repeat
-
     size_t tmp, x, lowIndex, element, subArray;
 
     if (array) 
@@ -21,15 +16,15 @@ void selection_sort(int *array, size_t size)
             lowIndex = element;
             for (x = size - subArray; x < size - 1; x++)
             {
-                // if checked value (x) is less than tmp, 
-                // make it the new lowest (tmp/index)
-                if (array[x] < tmp)
+                /** if checked value (x) is less than tmp, **/
+                /** make it the new lowest (tmp/index) **/
+                if ((size_t) array[x] < tmp)
                 {
                     tmp = array[x];
                     lowIndex = x;
                 }
                 
-                // at end of search, swap lowindex and init value
+                /** at end of search, swap lowindex and init value **/
                 if (x == size - 1)
                 {
                     array[lowIndex] = array[element];
